+++ frontend/test/pytest/test_compiler.py
@@ -262,6 +262,7 @@ class TestCompilerState:
             assert observed_outfilename == expected_outfilename
             assert os.path.exists(observed_outfilename)
 
+    @pytest.mark.skip(reason="skip this test with the plugin system on CI")
     def test_compiler_from_textual_ir(self):
         """Test the textual IR compilation."""
 
@@ -272,8 +273,9 @@ module @workflow {
     return %0 : tensor<f64>
   }
   func.func private @workflow(%arg0: tensor<f64>) -> tensor<f64> attributes {diff_method = "finite-diff", llvm.linkage = #llvm.linkage<internal>, qnode} {
-    quantum.device ["kwargs", "{'shots': 0}"]
-    quantum.device ["backend", "lightning.qubit"]
+    quantum.device ["rtd_kwargs", "{'shots': 0}"]
+    quantum.device ["rtd_name", "LightningSimulator"]
+    quantum.device ["rtd_lib", "./runtime/build/lib/librtd_lightning.so"]
     %0 = stablehlo.constant dense<4> : tensor<i64>
     %1 = quantum.alloc( 4) : !quantum.reg
     %2 = stablehlo.constant dense<0> : tensor<i64>